refactor(examples): 迁移 clipboard 至 plugins 目录
