+++ examples/src/App.tsx
@@ -33,7 +33,7 @@ import { CaseMindExample } from './pages/case/mind'
 import { CaseSwimlaneExample } from './pages/case/swimlane'
 import { OrgExample } from './pages/org'
 import { SnaplineExample } from './pages/plugins/snapline'
-import { ClipboardExample } from './pages/clipboard'
+import { ClipboardExample } from './pages/plugins/clipboard'
 import { KeyboardExample } from './pages/keyboard'
 import { DndExample } from './pages/plugins/dnd'
 import { ScrollerExample } from './pages/scroller'
@@ -88,12 +88,12 @@ function App() {
       <Route path="/case/mind" element={<CaseMindExample />} />
       <Route path="/case/swimlane" element={<CaseSwimlaneExample />} />
       <Route path="/org" element={<OrgExample />} />
-      <Route path="/clipboard" element={<ClipboardExample />} />
       <Route path="/keyboard" element={<KeyboardExample />} />
 
       {/* Plugins Example */}
       <Route path="/plugins/dnd" element={<DndExample />} />
       <Route path="/plugins/snapline" element={<SnaplineExample />} />
+      <Route path="/plugins/clipboard" element={<ClipboardExample />} />
 
       <Route path="/scroller" element={<ScrollerExample />} />
       <Route path="/selection" element={<SelectionExample />} />
